refactor: Splits up files
