+++ Sources/HaystackClient/Authentication/Authenticator.swift
@@ -1,0 +1,4 @@
+@available(macOS 10.15, *)
+protocol Authenticator {
+    func getAuthToken() async throws -> String
+}
