+++ Sources/HaystackClient/Authentication/ScramAuthenticator.swift
@@ -2,11 +2,6 @@ import CryptoKit
 import Foundation
 
 @available(macOS 13.0, *)
-protocol Authenticator {
-    func getAuthToken() async throws -> String
-}
-
-@available(macOS 13.0, *)
 struct ScramAuthenticator<Hash: HashFunction>: Authenticator {
     let url: URL
     let username: String
@@ -142,31 +137,5 @@ struct ScramAuthenticator<Hash: HashFunction>: Authenticator {
         case SecondResponseInconsistentHash
         case SecondResponseNoHeaderAuthenticationInfo
         case authFailedWithHttpCode(Int)
-    }
-}
-
-struct AuthMessage: CustomStringConvertible {
-    let scheme: String
-    let attributes: [String: String]
-    
-    var description: String {
-        // Unwrap is safe because attributes is immutable
-        "\(scheme) \(attributes.keys.sorted().map { "\($0)=\(attributes[$0]!)" }.joined(separator: ", "))"
-    }
-    
-    static func from(_ string: String) throws -> Self {
-        // Example input: "SCRAM hash=SHA-256, handshakeToken=aabbcc"
-        let scheme: String
-        let attributes: [String: String]
-        // If space exists then parse attributes as well.
-        if let spaceIndex = string.firstIndex(of: " ") {
-            scheme = String(string[..<spaceIndex]).trimmingCharacters(in: .whitespaces)
-            let attributesString = String(string[spaceIndex...]).trimmingCharacters(in: .whitespaces)
-            attributes = extractNameValuePairs(from: attributesString)
-        } else {
-            scheme = string
-            attributes = [:]
-        }
-        return Self(scheme: scheme, attributes: attributes)
     }
 }
