+++ Sources/HaystackClient/Authentication/ScramClient.swift
