+++ Sources/HaystackClient/Authentication/String+Base64.swift
