+++ Sources/HaystackClient/HisItem.swift
@@ -1,0 +1,11 @@
+import Haystack
+
+public struct HisItem {
+    let ts: DateTime
+    let val: any Val
+    
+    public init(ts: DateTime, val: any Val) {
+        self.ts = ts
+        self.val = val
+    }
+}
