+++ Sources/HaystackClient/HisReadRange.swift
@@ -1,0 +1,21 @@
+import Haystack
+
+public enum HisReadRange {
+    case today
+    case yesterday
+    case date(Haystack.Date)
+    case dateRange(from: Haystack.Date, to: Haystack.Date)
+    case dateTimeRange(from: DateTime, to: DateTime)
+    case after(DateTime)
+    
+    func toRequestString() -> String {
+        switch self {
+        case .today: return "today"
+        case .yesterday: return "yesterday"
+        case let .date(date): return "\(date.toZinc())"
+        case let .dateRange(fromDate, toDate): return "\(fromDate.toZinc()),\(toDate.toZinc())"
+        case let .dateTimeRange(fromDateTime, toDateTime): return "\(fromDateTime.toZinc()),\(toDateTime.toZinc())"
+        case let .after(dateTime): return "\(dateTime.toZinc())"
+        }
+    }
+}
